Refactoring Robot.cs To be in 13 different file.
Robot is divide using partial which won't change it's functionality.
